+++ src/Pages/InvitePage.tsx
@@ -1,9 +1,8 @@
 import { getSearchMember } from '@/apis';
 import { Check } from '@/assets/icons';
-import Button from '@/components/common/Button/Button';
-import Form from '@/components/common/Form';
-import IconWrapper from '@/components/common/IconWrapper';
-import { checkedStyle, checkIconStyle } from '@/components/common/Todo/TodoList';
+import { IconWrapper } from '@/components';
+import { Button, Form } from '@/components/common';
+import { checkedStyle, checkIconStyle } from '@/components/Todo';
 import { useToastHandler } from '@/hooks';
 import { useInviteMember } from '@/hooks/mutations';
 import { useSquadStore } from '@/stores';
