+++ src/components/common/Button/index.ts
@@ -1,0 +1,3 @@
+import Button from '@/components/common/Button/Button';
+
+export { Button };
