+++ src/components/common/Calendar/index.ts
@@ -1,3 +1,3 @@
 import Calendar from '@/components/common/Calendar/Calendar';
 
-export default Calendar;
+export { Calendar };
