+++ src/components/common/ErrorBoundary/index.ts
@@ -1,0 +1,3 @@
+import RootErrorFallback from '@/components/common/ErrorBoundary/RootErrorFallback';
+
+export { RootErrorFallback };
