+++ src/components/common/Form/index.ts
@@ -1,0 +1,3 @@
+import Form from '@/components/common/Form/Form';
+
+export { Form };
