+++ src/components/common/Loading/index.ts
@@ -1,0 +1,3 @@
+import Loading from '@/components/common/Loading/Loading';
+
+export { Loading };
