+++ src/components/common/Modal/ModalContents/ActionPrompt.tsx
@@ -1,29 +1,28 @@
-import Button from '@/components/common/Button/Button';
-import ModalButtonGroup from '@/components/common/Modal/ModalTemplate/ModalButtonGroup';
-import ModalContent from '@/components/common/Modal/ModalTemplate/ModalContent';
-import ModalTemplate from '@/components/common/Modal/ModalTemplate/ModalTemplate';
+import { Button } from '@/components/common';
+import { ModalButtonGroup, ModalContent, ModalTemplate } from '@/components/common/Modal';
+
 import { ActionModalContentProps } from '@/types';
 import { css, Theme } from '@emotion/react';
 
 const ActionPrompt = ({ onSubmit, onAbort, actionModal }: ActionModalContentProps) => (
-    <ModalTemplate isOverlay={true} onClose={onAbort}>
-      <ModalContent>
-        <p css={descStyle}>{actionModal.message}</p>
-        <ModalButtonGroup>
-          {actionModal.displayCancel && (
-            <Button text="닫기" onClick={() => onAbort()} name="cancel" variant="default" type="button" />
-          )}
-          <Button
-            text={actionModal.text}
-            name={actionModal.type}
-            variant={actionModal.type}
-            type="submit"
-            onClick={() => onSubmit(actionModal.text)}
-          />
-        </ModalButtonGroup>
-      </ModalContent>
-    </ModalTemplate>
-  );
+  <ModalTemplate isOverlay={true} onClose={onAbort}>
+    <ModalContent>
+      <p css={descStyle}>{actionModal.message}</p>
+      <ModalButtonGroup>
+        {actionModal.displayCancel && (
+          <Button text="닫기" onClick={() => onAbort()} name="cancel" variant="default" type="button" />
+        )}
+        <Button
+          text={actionModal.text}
+          name={actionModal.type}
+          variant={actionModal.type}
+          type="submit"
+          onClick={() => onSubmit(actionModal.text)}
+        />
+      </ModalButtonGroup>
+    </ModalContent>
+  </ModalTemplate>
+);
 
 export default ActionPrompt;
