+++ src/components/common/Modal/ModalTemplate/index.ts
@@ -1,0 +1,6 @@
+import ModalButtonGroup from '@/components/common/Modal/ModalTemplate/ModalButtonGroup';
+import ModalContent from '@/components/common/Modal/ModalTemplate/ModalContent';
+import ModalHeader from '@/components/common/Modal/ModalTemplate/ModalHeader';
+import ModalTemplate from '@/components/common/Modal/ModalTemplate/ModalTemplate';
+
+export { ModalButtonGroup, ModalContent, ModalHeader, ModalTemplate };
