+++ src/components/common/Modal/index.ts
@@ -1,2 +1,4 @@
 export { default as Modal } from './Modal';
+export * from './ModalContents';
 export { default as ModalPortal } from './ModalPortal';
+export * from './ModalTemplate';
