+++ src/components/index.ts
@@ -1,1 +1,5 @@
-export * from './layouts';
+import IconWrapper from '@/components/IconWrapper';
+import Notification from '@/components/Notification';
+import SquadIdGuard from '@/components/SquadIdGuard';
+
+export { IconWrapper, Notification, SquadIdGuard };
