+++ src/components/layouts/index.ts
@@ -1,5 +1,8 @@
-export { default as Container } from './Container';
-export { default as Footer } from './Footer';
-export { default as Header } from './Header';
-export { default as Main } from './Main';
-export { default as MainLayout } from './MainLayout';
+import Container from '@/components/layouts/Container';
+import Footer from '@/components/layouts/Footer';
+import Header from '@/components/layouts/Header';
+import LoginLayout from '@/components/layouts/LoginLayout';
+import Main from '@/components/layouts/Main';
+import MainLayout from '@/components/layouts/MainLayout';
+
+export { Container, Footer, Header, LoginLayout, Main, MainLayout };
